float

+++ foopackage/add.go
@@ -2,6 +2,6 @@ package foopackage
 
 
 // Add l,f
-func Add(a, b int) int {
+func Add(a, b float32) float32 {
 	return a + b
 }
